Add files via upload
Figured out how to send in a 'function' of sorts, so that eventually I can make a object structure that will allow an interface describing bindToParent, bindToChild, and raise warnings if attempting to uni-direction bind [Maybe figure out whether the attempt is a unidirectional binding and go forward with that].

So tomorrow an interface will be constructed, or a parent structure, to enforce attributes and methods for resolving those two object patterns.

Then we'll continue with adding to our layout structure by binding each child to it's parent [as it's more likely the closer to our leafs we are of our node tree, we will end up being bound to the child to fit [thus enforcing a min dimension, or a hide property]]. 

The event an object has a parent it is bound to, and a child it is bound to, there needs to be some thought given to whether there is precedence, or conditions where the binding must change the function it's to use. This suggests the need of a switch case anonymous function/object that we can push potential bindings to, and re-activate certain bindings for maximum flexibility... Or we could just be careful.