Add files via upload
Worked on interface implementations in Root. 
After bidirection's are tested in the testBindings file to get familiar with their behavior, the interface may be expanded.
Then a test file to explore a given scene tree will be created, this is in the event I need to trace a cyclic binding. Thus it may require an additional object or array to keep a copy of binding parameter references. I'm hoping there is an inbuilt method for detecting bound dependencies, but there may not be. 


[Preempt thoughts]
Bidirectional behavior needs to be tested before the interface entertains them.
A condition for debugging [to encourage faster layouts in the future] is a removable object or feature for tracking the bindings throughout a project. As the nodes are already in a tree, may create an optional 'English' translation of the bindings, and report them in a tree like fashion. However the only issues would likely be either 1) Bidirectional is not limited to 1 bind, or 2) multiple bindings created a cyclic condition somewhere, which can cause unexpected behavior due to feedback through the properties.

The bug this aims to pre-empt is a situation where you have 3 panes, A>B>C, where C is bound to A, and B is bound to A and C [min and max properties], but because C is bound to A, if A changes, then C changes, which changes B, but A and C may result in unexpected behavior in B (like expected it to shrink, but it was prevented from shrinking due to A].