Add files via upload
Part of the JH Paint project involves the canvas, thus as we know our ideal inner pane scaling within a tabPane due to earlier work [Rectangle bug investigation], we now simply need to draw items on the canvas, and implement some features regarding those. This will be somewhat lengthy as the documentation leaves much to be desired in explanations and examples.

gc.arc doesn't seem to work, but gc.fillArc and other derivatives work well enough.
This folder contains some objects and tools to be used for testing. The button will cycle through user made strings to draw a variety of options from the switch statement, to get a 'feel' for the graphicContext library. Two methods are currently provided 'First Stroke', which is just where I'll be testing options, and 'Reset', to reset the canvas.

Later today I intend to investigate the other premade functions in the library, and then look at SVG Paths.

I currently don't have a data structure I'm enthusiastic about for Undoing, Redoing, and Drag/Drop strokes and shapes. It is imperative that a undo/redo tree is able to occur, but dragging and dropping, or being able to reselect shapes dropped would be nice. Or a feature to specify the exact length and coordinates of the shape.