Add files via upload
Unexpected behavior occurred. Still investigating the methods to see the 'proper' usage. Eventually the idea is to make interactable objects that then get replaced with the UI points/coloring before being drawn.  

The main is the function for testing between different draws and seeing if there are artifacts left behind, which may explain the unexpected behavior.

Some progress was made in learning the methods, but some interesting behavior was found that'd be nice to snapshot.

A drag and drop feature would require learning more about Java threading and the underlying assumptions with the Canvas object. So the UI being a separate component would be nice. Before the drag and drop also, creating it so that a selected region 'breaks' the elements, or layers different objects to hide the break would be nice too.