Update README.txt
Updated some thoughts and technical set up details to be in line with the April Restart. 
The April Restart was spawned by a Talent Company desiring a Java candidate, so I rotated to this project again. Having some experience in JavaFX and more experience in Javascript and React, I am reorganizing the project and focusing more on subgoals for learning JavaFX with the Paint project guiding some of the inquires.