Add files via upload
Additional tests were made in the switch statements in CanvasTestObjects.Main.java, while there are observations still to be made with effect, and with local/global blendMode combinations, and with custom blends and effects, I figured it'd be best to start getting a system to draw and drag multiple things, and ensure that the blend modes can interact properly when separated by panes. They *should*, but getting a test up is good.

Additionally, in CanvasDrawUI, where the test will take place, the subgoal to test and create before the above is being able to draw a shape using nodes, and then 'lock' it into place. This involves at least 3 layers/panes for the desired scaling and coloring effect. The Controls object is a better more accurate version of the testObject using bindings to ensure values are properly synched. 

Later tonight will continue to work on the testObject, particularly with dragging objects, creating a polygon/line using those objects, and using a textInput box to spawn the nodes.

Some logic will need to be edited once zoom scale is introduced, as while the inner panes [canvas's] can be properly scaled, controlling the draggable nodes, and auto scrolling, and the proper bindings may need to be adjusted.